Don't assume the only Dying machines are currently losing votes.
If you:

```
  $ juju bootstrap
  $ juju enable-ha
  $ juju remove-machine 2
```

At this point, assuming machine 0 is the primary, then machine 1 will
also have lost its vote to preserve the odd-number of voters.
However, if you then
```
  $ juju remove-machine 1
```

Then that machine will go into Dying, and *should* lose `JobManageModel`
but it wouldn't, because it was not losing its vote.
Instead we just check all machines, see if they are not Alive, and flag
anything that has finally lost its vote as no longer managine the model.