Don't remove non-voting members in peergrouper
The members map returned from peergrouper.pgWorker.updateReplicaSet is
only used to get addresses of API servers to publish to the hub, and
since even the non-voting members are running an API server they should
be included. This fixes problem where removing a controller machine from
a 3-machine cluster results in one of the raft workers in the remaining
machines not starting up.