Change the settingsMap to encode correctly.
Implement a custom GetBSON so that settingsMap properly escapes its
content when being serialized. That helps so if we do access the
documents directly, we'll always decode on read and encode on write.

Update the upgrade test so it includes being able to modify a doc that
has uncommon characters.