Push further on removing unneeded code.
Since we no longer flag controllers for removal, we don't need to have a
'remove' or 'demote' fields on enableHAIntentions. And then we can drop
all the related functionality.

removeControllerOps becomes safe on concurrent updates to the list of
controllers, and we re-use it via RemoveControllerMachine instead of
doing all the ops directly in a single transaction (instead we can use
Run() and have it give nice errors if something really is wrong.)