There was a data race in Refresh() vs SetHasVote()
Refresh() was setting fakeMachine.doc as an attribute, and SetHasVote
was reading m.doc.id.
Now we get rid of the local doc object entirely. This invalidates the
use case of Refresh(), which might be wrong, as real Machine objects can
have in-memory state that doesn't match the backing DB state. But the
Set* functions mutate in-memory and backing state. They can only
disagree because there may be a different process writing to backing
state.