Support removing a primary when in N=2 mode.
If we are down to N=2, we need to restore the secondary as a voter
before we can ask the current primary to step down.
Given how long it will take us to re-evaluate the loop, we might want a
faster polling when we are in these transitionary phases.
But this should at least be the correct transition steps.