Provide status time to all outputs
The following change adds the status-time to all outputs that are
in the long form (json, yaml, tabular).

Most of the work includes changing the tests to acknowledge the new
field. Including the more generic approach of substituting the fake
time stamp from the since one, to the now parameterised key.

One thing I'm not 100% settled on and open for questions is the
following:

 1. Naming of the field
 2. Should the field be nested inside a object. Think `meta.status-time`
 3. Should we gain the status from the API instead of using time.Now()