Introduce a marker interface which is embeded in commands which can only run on IAAS models. The base model command will attempt to determine model type as early as possible, either in Init() or Run(). Some commands don't operate on a default model and so the model type can only be determined in Run(). Once the model type is known, any IAAS only commands are rejected.
Because model type is attempted to be looked up in Init(), some refactoring of the base command was needed, eg we now get the client store in Init() whereas previous that was disallowed (for no real reason).

Also, a number of tests will need updating as they aren't properly creating a client store to use if only Init() is invoked (but we now need it).