Controller Timestamp
The following commit introduces a controller timestamp. Instead of
using local time of the client CLI, we call the controller via the
FullStatus RPC call. The call adds a ControllerTimestamp using the
model clock as the source of the time.

If no ControllerTimestamp is found (controller is an older variant)
the we don't print the timestamp in the status output.