Add the raft-backstop worker
This will recover the raft cluster by force-appending an updated
configuration if we remove the leader in a 2-node cluster.

The raft-worker will bounce after a minute with no leader contact, to
let it see the configuration change.

Since it's not permitted to open a boltdb store twice, we expose the log
store as an output of the raft worker so that the backstop worker can
use it to append an updated configuration if needed.