Gometalinter output parser
The following bash script validates the output from gometalinter to
make sure that it's a) valid and b) is not a mock file.

The output for the running script is a tmp file. If you want to debug
the script, then changing `OUTPUT_FILE=` to point to a local output.