Start working on better netplan support.
We were not only missing Bonds but also VLAN which will be a pretty big
issue.
Also, while digging through the code I found a bunch of attributes that
we weren't supporting, and decided to grab the netplan example files and
bring them in as tests.
A couple of the tests also don't run on Windows so let's skip those.