Now we support a lot of the Bond parameters.
We can handle if a given parameter is in the form of an integer, or if
the parameter is in the form of a string. goyaml would let us
deserialize an integer into a string value, but we want to be able to
round-trip what was already in that field without introducing extra
quoting.