Start checking that we can round-trip properties.
We haven't gotten everything working, but it is a step in the right
direction.
Note that 'dhcp4:' *defaults* to True, which means you can't just use
'omitempty' with a boolean, because *disabling* it will cause it to be
immediately re-enabled as we omit the field. But using a *bool should do
the trick.
Also 'match:' can be omitted for ethernets.

We still need VLAN support for us to handle round tripping all the
examples.