We can now round trip the example files.
Updated attributes for routes, routing-policy, vlans, etc.
Only problem file was testdata/examples/direct_connect_gateway.yaml
However, that failed because it wasn't a list of ethernets, but only a
top-level configuration. I'm not sure if it was properly formed, but if
it was, we have to support a very different way of parsing.