Fix race in raftclusterer tests
The race was caused by resetting s.reqs to a new channel in the test
setup when a subscribe callback had previously read s.reqs from a
different goroutine (in a previous test). Avoid it by having the
callback send to a local variable so that setting s.reqs in a subsequent
call to SetUpTest doesn't overwrite the old channel.