Turn off ShutdownOnRemove in test fixture raft config
With this on, the leader raft node shuts down after being
demoted. Generally, the test then manages to finish ok because it does
before another node takes over, but occasionally one of the other nodes
reaches their heartbeat timeout and gets elected. In that case it seems
to go into a state where it can't be shut down cleanly because it can't
talk to the former leader (which has itself shut down) so the teardown
hangs.

This seems to be a bug in hashicorp/raft (or at least the version we're
using) - I'm still chasing it down, but this fixes the hang in our tests
anyway.