Two changes, FAN IPv6 and Openstack overlapping subnets.
When autodetecting FAN overlays, we were taking all SuperSubnets and
mapping them, but we know that we don't want overlays for IPv6 networks.
Openstack SuperSubnets always returns all subnets, and it isn't clear if
SuperSubnets should be filtering v6.

For Openstack subnet listings, we now only list the subnets that are
part of either the 'network' config or the 'external-network' config,
rather than listing all networks all the time. (This should address
bug #1733266 until we have support for Openstack instances spanning
multiple networks.)