Improve the UX of the timestamp
Following feedback, changes to the location of the controller
timestamp location has changed, to be inline with the model table.

Hopefully the controller timestamp with the context of the table
makes it easier to understand what it relates too, rather than a
dangling orphan.

The following changes change where the status text is displayed. It
moves the text from the bottom of the output, to next to the model/
controller output. This removes the dangling orphan issue.

 - Make sure we test the time