Lock gometalinter as a dependency
The following commit locks the gometalinter as a dependency, so any
upstream changes to the gometalinter don't potentially break the
juju release process. In order to do this, we have to force the
gometalinter into the juju repository. We do this by making a test
that checks what the output of the bash file is looking for. This
should protect us when upgrading gometalinter in the future, by
erroring out if there are differences.