EnableHTTPSListener now resides on the lxd server
The following code moves the EnableHTTPSListener to the lxd server
code. Most of the code in tools/lxdclient/utils was deleted and the
code left in there is just for test setups.

The provider now correctly asks the same lxd server for the same
information as other calls surrounding it, making a lot more
consistent