Use new local server for access
The following changes how we retrieve the local host address. The
major change is now we have to VerifyNetworkDevice so that we can
access the LocalBridgeName (otherwise it's empty).

We then can reuse the same code base for quick access for Zenial
and then slower access for Bionic as that requires retrying...

Also remove the rawProvider interface code, so that we're not
bloating that out.

Backed off the delay, to 2 seconds.