Check the timestamp correctly
We now check for the format before sending it to the parser. Some of
the tests send the information without a Z inside of them and can cause
issues for validating the time.Parse. We check for this and make sure
we insert the missing information correctly. This fix should work for
both iso and non-iso time.