Improved error messaging around file locking
The current implementation doesn't distinguish between a file lock
problem and a problem with the reading/writing (think input/output)
of the standard flow.

To help with this, we tell people that there is a problem with a
lock file, by giving them much better error messages. That way they
maybe able to resolve the issue themselves.