Provider Credentials should use mocks
The following provider credentials should use a cert read writer to
access files on disk. When constructing the provider credentials in
the provider, that should provide a stdlib implementation. Then
during testing, we can use mocks to better define side effects.

This makes the code more explicit and less dependant on file systems
to make sure files are in the right place at the right time.