Use container/lxd
The following commit removes the tools/lxdclient from the provider.
The tools/lxdclient is around until we move all the storage checking
required, before we cull the package.

We also simplify the creation of the server, where by we pass a
factory around so that mocks can utilise it. Along with cleaning up
what's local and what is remote communication.