Add raft FSM for storing lease information
This is intended to replace the mongo-stored leases. Stores leases in
memory, allows updates through the Apply method and handles snapshotting
and restoring.

Updated yaml dependency so we can use En-/Decoder.

Add raft forwarder worker - this forwards lease updates to the leader
raft node and replies with the results.

Move lease secretaries from state to worker/lease so the lease worker
manifold can use them too.

Add manifold for the lease manager so it can run in the dependency
engine. It can't live in worker/lease because the manifold uses state,
and state depends on worker/lease (since the mongo-based lease store
manager runs as a state worker).

Add lease-manager to machine manifolds - changed the raft FSM to be
the core/raftlease one. (We'll need to handle the case where a
simple-fsm snapshot is restored into a raftlease FSM.)

Extend globalclockupdater to work with the raft lease store. The lease
store implements globalclock.Updater, and the lease manager manifold
exposes it as output. Change the manifold so that it can depend either
on state or on the lease-manager, and then add a new instance of the
clock updater to the machine agent manifolds.

Make the apiserver depend on the lease manager, and thread that through
to the facade context. Make the API facades use that instead of getting
the checkers and claimers directly from state. Add legacy-leases feature
flag to control whether the leases go through raft or state.

Pass logger into globalclockupdater, otherwise we can't distinguish
between the logging for the two running updaters.

Update State.ApplicationLeaders to work with raft leases - unless the
feature flag is set it reads the leaders from the leaseholders
collection instead.

Updated manifold tests with new workers - lease-clock-updater,
lease-manager and raft-forwarder.

Also tweaked AssertManifoldsDependencies so that it reports what names
are missing or unexpected.

Implement a dummy lease store for the dummy provider, and update tests
that claim leases to use the dummy lease manager

Set legacy-leases for the cmd/jujud/agent tests - these tests fail
without this flag set at the moment. I'm fairly sure it's due to the startup
issues I'm intending to fix next, so I'm setting the flag to keep these
tests passing now, and will fix them properly in the next step.