Changed core/lease.Client methods to take lease key
This will allow the lease manager to be reworked to have one worker
running for all the singular and leadership leases in the system, rather
than a singular and a leadership worker running in each State.

For the existing state client implementation, since it knows it's got a
specific namespace (from config) and model (implicitly through the mongo
connection), it ignores everything except the lease name.