Local Server Factory changes
The following commit ensures that the server creates a new local
server, that has all the connections setup correctly, before handing
it back. The reason for this is so that we don't have to place
retries everywhere outside of the server initialisation code path.

This should ensure any consumer of the factory can then ensure that
when they get a local server, they are in fact getting the right
one.