Clean up the ServerFactory
The following cleans the server factory up, in doing so, make it a
lot easier to understand how to create the server factory correctly.

We still create eschew the lxd container functions so that it's
easier to mock, but at least these are now encapsulated a way from
the consumer of the factory.