Check for supported versions
The following checks for the servers supported version, before it
is being used by the provider. It follows the old semantics of not
checking for simple streams.

I'm unsure if this is in the correct place to check, or indeed if
we should be checking the type instead of using the connection info,
but I thought it would be a good PR start