Removing critical error
The following removes the critical error, I'm not sure it should
be there, because we return the error directly under the log
statement.

An example of this error showing in the wrong place is when you've
bootstrapped to lxd on localhost and deploy something. When you
try and grab some metrics about that deploy and you miss type a
model name, it comes with a critical log.

```sh
juju bootstrap localhost lxd-test
juju deploy postgresql
juju metrics --all -m=lxc-test
```