Fix for LP 1777897
Allow the OpenStack client to use a ca cert for communication with the cloud.
This does not apply to the cinder client and ca cert use with the juju
storage provider functionality.

Change the juju add-cloud OpenStack url verification to use a client
which does not do ssl hostname validation.  This allows the url to be
validated even if a ca cert is required.

Drive by change of strings to constants in openstack provider config.go