Sort API Host Ports by server too
Currently the API Servers are sorted by HostPort for each server
(ie, inner slice) but not by server (ie, outer slice).

This allows for unnecessary updates of API servers (and triggers
of rsyslog restarts, for example) in case the servers and their
hostports have _not_ changed but the _order_ of servers changed.

It happens because publishAPIServers()'s apiServers parameter is
obtained from pgWorker.apiPublishInfo(), which iterates over the
pgWorker.machines _map_ with 'for/range'.

However, the iteration order over maps is _not_ specified and
_not_ guaranteed to be the same from one iteration to the next
(Go spec, section 'For statements', subsection 'For statement
 with range clause', note 3 [1]); apparently the runtime even
randomizes the map iteration order.

So, sort the API servers host ports by server as well, so that
the order of the servers does not change if the servers do not
change. This uses String sort on the host ports of each server.

[1] https://golang.org/ref/spec#For_statements

Signed-off-by: Mauricio Faria de Oliveira <mfo@canonical.com>