Bump the version from Xenial to Bionic
This makes sure that all LXD images by default are Bionic for 2.5
release. The issue with this is that it can cause issues from this,
because we're now bumping the LTS version for EVERYTHING.

A lot of testing should be done to ensure that we don't cause other
issues for other parts of the system.