Fix the following tests around providers
The following updates to the test code updates to the bionic LTS
changes. Most of the code is in export_test, this isn't ideal in
fact I believe this should be moved to a more appropriate location
i.e. package_test.go. export_test.go files should be used to
re-export types or functions.