Default Constraints Checker
The following commit ensures that when we bootstrap, we only apply
the default constraints for providers that require it. This is an
opt-out method and not an opt-in and should be backwards compatible
for all existing providers.

The reason for this change is to allow some providers (think LXD),
to opt out of using default constraints, without inserting magic
numbers/strings into the constraints vocabulary/logic, which could
cause potential bugs down the line. With this method of change, we
are asking if the Environ is a type and if the environ implements
the type, asking if we should use the default constraints. This
logic is simple and doesn't tie any provider to the bootstrap
directly and feels less hacky than other ideas floating around.

If there are better ways to solve this, then I'm more than happy
to change this.