Default lxd region
Update the pollster interaction so that we can have default naming
for object key names. To improve the UI/UX for adding a LXD cloud
interactively we want to state that there can be a region and that
it can be default, we do allow you to override this when adding a
cloud, but a sane default should suffice.