Openstack default fix
The following also checks for empty default string, so that we're
backwards compatible with other providers like openstack. Openstack
here adds a default of "" and if a provider sets the default to ""
we skip it, like it was equal to nil.

Also updated to include drive by comment removal and better default
wording for LXD provider.