Register Credentials for Providers
The following change is a fundamental change for builtin providers
like LXD. With this change each provider (opt-in) can suggest some
potential credentials that can be stored in the credentials yaml.

This makes LXD less special, so all subsequent calls to the cloud
will always find a predefined localhost credential with in the
credentials yaml.

The main driver to this is to allow multiple LXD clouds and each
one can then define their own credentials without fearing the
multiple credential error when interacting with LXD.