Implement netplan merging in our code
This was done using yaml.UnmarshalStrict to cumulatively update a
Netplan struct, but the new version of UnmarshalStrict refuses to
overwrite map keys that already exist in the target. We're relying on
the behaviour of strict to ensure files with unknown attributes don't
get loaded, so we load into a local Netplan value and then copy values
from the local one to the target.