Provider Names
The provider should pass back the name of the credentials it's
registering. The cloud name when bootstrapping is irrelevant when
the provider is registering the credentials. As these are built-in
default credentials and may not have any ties to the one you're
trying to bootstrap.

So in this instance we pass back a map[string]*cloud.CloudCredential
for the bootstrapping code to know that the string in the map
relates to the credential cloud name. In practice the key/cloud name
is "localhost" for LXD. For other clouds this could be something else.

 - go generate