Don't crash the lease manager on timeouts
At startup there's a period where the lease manager is running but raft
isn't yet. We get timeouts when workers try to claim singular leases,
but if the lease manager bounces because of them it causes a lot of
fallout because that also bounces the API server, breaking API
connections and invalidating watchers.

Instead we log the timeout error, as well as responding to the client in
the case of claims/extensions timing out.