Increase the raft-forward timeout threshold to 5s
It needs to be high enough that when we see it we can be very confident
that it happened because there was no raft-forwarder listening on the
hub. The old value was low enough that we'd sometimes see false reports
of timeouts even though the request was actually received and applied to
raft. This would mean that the notify target wouldn't be sent the update
if a lease was claimed or expired, for example.

There's not much downside to having a high-ish value - the caller will
be blocked waiting for a response, but other lease operations aren't
queued behind this one.

Also update the pubsub dependency so we don't see debug logging every
time a clock update is sent.