Allow OpenPorts/ClosePorts if we don't have an external network.
We just treat them as no-op, since we don't have any firewalls to
actually apply. Arguably users shouldn't be exposing in places that
don't have an external network, but they wouldn't be able to unexpose
(and thus close the ports) unless we treat both Open and Close as a
no-op.