Add modelUUID to facade context LeadershipClaimer
This will be used by the migrationtarget facade when importing a model,
so that we can claim leadership for the right units once the model's
imported.

We can only support getting a claimer for a different model when using
raft leases. Allowing it for legacy leases would require getting the
state from the state pool, which would need a way to close the
leadership claimer so the state could be released. Instead the import
code will check the feature flag and use the state claimer directly.