Claim leadership in migration.Import
It couldn't be done purely in state.Import with raft leases,
since the lease manager now runs in the dependency engine and isn't
accessible from state. Instead the migrationtarget passes in the
leadership claimer function, and once we have the uuid from the
imported model we use that to get the right claimer.

The facade context state is for the controller model (since this facade
call will create the imported model), so we can't use that same claimer
when we're using legacy leases. We *could* get the imported model state
from the facade context state pool, but we'd need to find some way to
release it at the right time - either by adding Close() to
leadership.Claimer, or making facade.Context.LeadershipClaimer return a
closer func. It seemed simpler to just make the import code look at the
feature flag and get the claimer differently based on that.