Optional schema values - hide from UI
The following commit attempts to hide certain values from the UI,
whilst still providing values for the schema at a later stage. The
idea for this is that LXD allows the usage of a trust-password, but
that is transient as it's replaced with certificates as soon as
a connection to the server is made.

So the schema really needs to only show a trust password, then the
following values are then filled when a communication is setup. It
seems rather backwards and unique for LXD, the reason being that
the schema that we're using is rather rigid and provides no current
flexibility to enable this sort of flexibility.

It should be noted that the reason we don't keep the trust-password
and just re-establish every time using the trust-password is because
the LXD security documentation[1], categorically states that the
trust-password should be turned off once all server have joined the
cluster.

[1] https://github.com/lxc/lxd/blob/master/doc/security.md#production-setup