Back out HideFromUI
So with the following change we no longer need HideFromUI, as we
convert all other auth types into certificate auth type. So we
only need to add a trust-password for the schema and once we require
connectivity to the lxd server we get the certificates and fill them
into the credentials yaml file and we no longer need to talk about
interactive mode.