Split Trapdoor out of raftlease.NotifyTarget
This is a first step for moving change notification out of the lease
store and into the raft forwarder. The notify target will be held by the
forwarder which can say definitively whether a command has been applied
to the FSM, which the store can't do because it's on the other side of
the hub. The trapdoor func will stay on the store.