Mock MeterStatusAPI
The following mocks out the meter status api. The tests mock out
the state type and consume only the methods that we require. The
new tests don't require the database when running "WatchMeterStatus".

Interestingly we require a new constructor so that we satisfy the
allfacades registry setup, as that actually tests the function
type is (*state.State, facade.Resources, facade.Authorizer),
which is a shame, because that restricts what we can get the meter
status to consume.