Fix intermittent failure in TestRemoveApplicationStopsWatchingApplication
In some cases, the worker loop hadn't had time to get the life for the
gitlab before the test would set the notfound error - in that case
it would never start the watcher for gitlab, and never kill it.

Fix this by setting errors (starting with a nil) before firing any
change events.