Fix a race in state watcher makeControllerIdFilter
The returned filter closure updates a closed-over variable to keep track
of the last machines seen (for cases when st.ControllerInfo fails - I'm
not sure under what circumstances that happens but it seems important to
maintain that behaviour). The function can be called from multiple
goroutines so the state update causes a data race. This *might* be the
cause of an intermittent failure in featuretests.upgraderSuite.

Rather than adding a lock to the closure, it seemed clearer to rewrite
the filter as a struct with a match method.