Fix race in caasoperatorprovisioner test
It would wait until there was at least one call on the caasClient stub,
but then check for two calls. Most of the time the two calls would
happen in quick enough succession that the test would pass, but
sometimes (especially under load), it wouldn't. Also add an assertion on
the Calls() length to avoid the test panicking (since the CheckCallNames
doesn't stop it).