Ensure target is defined firstly.
The following patch is required when attempting to bootstrap into
a lxd cluster, but the container image isn't found. We believe that
the happens because of an optimisation[1] with in lxd which assumes
the instance is the same as the source. Because of the way the methods
are invoked in juju, can lead to the prior bug[2].

[1] https://github.com/lxc/lxd/blob/c3de74dc45c23c177328a8d7c2a197ebcd521b42/client/lxd_containers.go#L176
[2] https://bugs.launchpad.net/juju/+bug/1793291