Allow passing force
The following passes force as an option to the deployment info, so
that we can verify if we need to show an error or not when
validating a deploy.

The force flag overrides the error, but still validates the charm,
with the aim to improve error messages when attempting to debug
a deployment from the logs.