Add controller API port config option
This will allow controller connections to happen on a separate port to
other agent connections, and the standard API port won't accept
connections until the controllers are connected. Also add an option to
further delay accepting connections for some time to allow the
controllers to stabilise (for upgrading/restarting very busy
controllers).

New controller config attribute, and writes to agent.conf.

Add api-port-open-delay controller config value.

Add the controller-port worker and have the peer grouper and http server
depend on it. Also fix the shutdown of the http-server so it blocks new
accepts on shutdown.

Make sure the httpserver and peergrouper require the controllerport
worker to have been started.