Extract lxd profile content
The following is a very much WIP (version 3 - sigh), that aims
to get the lxd profile content out as fast as possible so that we
don't slow down the status output.

The following assumes that the serialised name of the lxd profile
is rebuildable on the application status side. There could be times
when this might not happen, but later commits should improve the
reliability of this.