Retrieve LXD profile names for containers
The following commit attempts to get profile names from the LXD
server API. At the moment the LXD profiles are filtered to remove
default names from the raw source. All the names that are returned
from the server are kept in the same order as returned from the LXD
API, what is currently missing is ensuring that we don't add any
potentially invalid information to the database that might not
actually be valid for the applications living on the machine.