Get all handlers registered before httpserver starts
Having api-server and raft-transport (which register handlers) depend on
the http-server so that they can get the mux and authenticator means
that there's a time between when http-server starts and those workers
start where we respond with 404s to valid API requests.

Instead, create the mux and authenticator in their own
worker (http-server-args, for want of a better name), have api-server
and raft-transport depend on that, then have http-server depend on both
of those so that it doesn't start until they've registered handlers.

This required tweaking raft-transport so that it registered the handlers
before returning the worker. Also needed to move the ifFullyUpgraded
gate to the raft worker, because otherwise there's a dep cycle from
raft-transport -> upgrader -> api-caller (implicitly) -> http-server ->
raft-transport. Moving it's fine because the raft transport doesn't need
anything that could be changed in an upgrade step (although raft-worker
does).

Tests not updated yet - committing this so I can apply to 2.5 and test
upgrading.