caasoperator test fix
Encountered a panic in assertUniterStarted running under stress-race:
the uniterStarted count was incremented before the channel was stored,
so in rare cases the attempt loop could get a null from
applicationChannel. One way to fix this would be to do the increment
after the store, but instead I rewrote it using channels (since
uniterStarted wasn't used for anything).