Rework the api-caller worker retry strategies.
This changes how the workers try to connect to the controller, and how
they retry their connections.

Rather than trying all controllers every 200ms for 3s on, 3s off, this
relies on the exponential backoff of the dependency engine. And instead
we try 1 controller at a time, with a 200ms delay between controllers.

We also shuffle the order that we try controllers with each retry. This
should help balance the load on controllers, and avoid having them
hammered when starting up.

We max out the delay between retries to 2 minutes, and backoff by 20%
each time. This means it takes around 20 retries in 10 minutes before we
cap out at 2min per retry. (which is way below the original 1500 retries
in 10 minutes.)