Fix cmd/jujud/agent tests
These were broken because the tests tried to change the port the
httpserver listens on by changing the agent config file, but it's now
determined by the controller config. This meant that the httpserver
wouldn't start because the port set in the controller config was already
claimed by the test httpserver started by the dummy provider in
JujuConnSuite. Update the controller config (by allowing api-port to be
changed) with the new port that's being written to the config, so the
httpserver can start.