When looking up IP Addresses we would block.
Because we were accessing addresses in any order, it would cause the
whole test suite to hang because IPAddrResolver would never return, and
that is done in the main loop before we spawn goroutines to call the
various addresses. It may be that we need to fix the code, but for now
we just fix the tests.