Use platform independent namespaces for app name
The previous code would fail if we use "names.Juju" as that is
different between windows and linux. The former having the ".exe"
suffix.

Unfortunately the way that the client status builds up the charm
profiles to show inside the status output, requires the name.Juju
to build the path for location. Unfortunately we built the names
of the charms on linux and tried to identify them on windows, the
missmatch would then be:

 linux:   "juju-controller-lxd-profile-0"
 windows: "juju.exe-controller-lxd-profile-0"

The match would never have and would cause the status to not output
the right field for windows.