Independent charm instance information
The following changes move the charm data information into a
independent collection. This should improve the number of watchers
that become alive when something changes on a machine. Although
an optimisation without measurement, it seem reasonable because of
how many watchers watch for change on the machine and become awake
might large deployments.