Change uniter to track config hashes
RemoteStateWatcher now uses unit.WatchConfigSettingsHash and stores the
hashes that the watcher returns.

ConfigHash, TrustHash and AddressesHash are stored on operation.State so
that they are persisted into the uniter state file.

The resolver compares the hashes on LocalState and Snapshot to determine
whether config-changed needs to be run.

For now, config-changed will still be run every time the agent starts
because the RemoteStateWatcher still increments ConfigVersion when the
trust config and addresses watcher signal, rather than storing
hashes. These will be swapped out once the hash versions of the
watchers are added. Then ConfigVersion can be removed from LocalState
and Snapshot and the hook will only run when one of
config/trust/addresses has actually changed.