Use Unit.WatchAddressesHash in RemoteStateWatcher
Remove ConfigVersion from LocalState and Snapshot - this is now
supplanted by Config-, Trust- and AddressesHash.

Fix the uniter tests that expect the config-changed hook to fire after
the worker bounces - this doesn't happen now that the hashes are
persisted in the uniter's state file.

UniterSuite.TestRunCommand panics for some reason - I haven't worked
that out yet.