apiserver/facades/client/cloud: avoid naturalsort
The tests uses naturalsort to sort UUIDs which is
a bit weird and means it's harder for external packages
to test against. Avoid that, and also make the tests avoid
the ordering dependency by putting the UUID into the
mock model (in general an API endpoint implementation
should be free to do its work in an arbitrary order or
even concurrently if desired).