Ensure we handle situations where profile is empty
The following checks to see if the profile might be empty when
getting the information about the charm. As charm profiles can
sometimes be nil, that check is encoded into the codebase
throughout, except here.

A drive by is also added to the status output to ensure that we
also don't print nil if the upgrade happens and there is no lxd
profile, yet there is a placeholder for it.