Error out if we're in a error state
The following should ensure that we don't allow the resolver to
pass if the lxd profile is in an error state and we can not pass
because the machine we promised is not available.

Additionally added unit tests to the both upgrade charm profile
AND upgrade series, so that we know what the resolvers are
attempting to do.